changing to doubles so that user does not have to add 'f' for a float

+++ week2/change_converter/app/src/main/java/change_converter/ChangeGenerator.java
@@ -7,7 +7,7 @@ import java.util.ArrayList;
 
 public class ChangeGenerator {
 
-  public static ArrayList<String> convert(Float value) {
+  public static ArrayList<String> convert(Double value) {
     String formattedValue;
     ArrayList<String> notesAndCoins = new ArrayList<>();
